+++ lib/ncs_navigator/warehouse/transformers.rb
@@ -13,6 +13,7 @@ module NcsNavigator::Warehouse
   module Transformers
     autoload :Database,        'ncs_navigator/warehouse/transformers/database'
     autoload :EnumTransformer, 'ncs_navigator/warehouse/transformers/enum_transformer'
+    autoload :SamplingUnits,   'ncs_navigator/warehouse/transformers/sampling_units'
     autoload :VdrXml,          'ncs_navigator/warehouse/transformers/vdr_xml'
   end
 end